move nilearn.plotting.matrix_plotting under nilearn.plotting.matrix
