+++ nilearn/plotting/__init__.py
@@ -32,7 +32,7 @@ from nilearn.plotting.img_plotting import (
     plot_stat_map,
     show,
 )
-from nilearn.plotting.matrix_plotting import (
+from nilearn.plotting.matrix import (
     plot_contrast_matrix,
     plot_design_matrix,
     plot_design_matrix_correlation,
